✨ feat(reviewer_submit_view.dart): add ReviewerKarakterSection widget and subtitleStyle method The ReviewerKarakterSection widget is added to the ReviewerSubmitView to display the character of the debtor. The subtitleStyle method is also added to provide a consistent subtitle style throughout the view. The KarakterCard widget is removed as it is replaced by the ReviewerKarakterSection widget.
🎨 style: refactor code to improve consistency and readability
This commit refactors the code in several files to improve consistency and readability. Specifically, it does the following:

- Adds a required `subtitleStyle` parameter to the `ResultInputSection` and `AnalysResponse` widgets to improve consistency in the use of text styles.
- Removes the `subtitleStyle()` function from `ReviewerSubmitKeuanganButton` and `AnalysResponse` and replaces it with the `subtitleStyle` parameter to improve consistency in the use of text styles.
- Adds an `elevation` of 6 to the `Card` widget in `ReviewerSubmitKeuanganButton` and `ResultInputSection` to improve consistency in the use of elevations.

🎨 style(reviewer_submit): refactor ReviewerKarakterSection and ReviewerKeuanganSection
The code has been refactored to improve readability and consistency. The ReviewerKarakterSection and ReviewerKeuanganSection widgets have been updated to include a subtitleStyle parameter, which is used to style the subtitle text. The ReviewerKeuanganSection widget has also been updated to include an elevation property for the Card widget.
