+++ lib/app/modules/reviewer_submit/views/reviewer_submit_view.dart
@@ -4,6 +4,7 @@ import 'dart:convert';
 // 🐦 Flutter imports:
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_analys_response.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_inputan.dart';
+import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_karakter.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_keuangan.dart';
 import 'package:flutter/material.dart';
 
@@ -112,6 +113,13 @@ class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
     return const TextStyle(
       fontSize: 20,
       color: secondaryColor,
+    );
+  }
+
+  TextStyle subtitleStyle() {
+    return TextStyle(
+      fontSize: 18,
+      color: Colors.grey[600],
     );
   }
 
@@ -227,10 +235,12 @@ class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
                       iconDone: iconDone(),
                       iconNotYet: iconNotYet(),
                       buttonStyle: buttonStyle(),
+                      subtitleStyle: subtitleStyle(),
                     ),
                     const SizedBox(height: 20),
                     AnalysResponse(
                       controller: controller,
+                      subtitleStyle: subtitleStyle(),
                     ),
                     const SizedBox(height: 20),
                     ReviewerKeuanganSection(
@@ -238,9 +248,13 @@ class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
                       iconDone: iconDone(),
                       iconNotYet: iconNotYet(),
                       buttonStyle: buttonStyle(),
+                      subtitleStyle: subtitleStyle(),
                     ),
                     const SizedBox(height: 20),
-                    KarakterCard(controller: controller),
+                    ReviewerKarakterSection(
+                      controller: controller,
+                      subtitleStyle: subtitleStyle(),
+                    ),
                     const SizedBox(height: 20),
                     BisnisCard(controller: controller),
                     const SizedBox(height: 20),
@@ -820,115 +834,6 @@ class BisnisCard extends StatelessWidget {
                 ],
                 validator: FormBuilderValidators.required(),
               )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class KarakterCard extends StatelessWidget {
-  const KarakterCard({
-    Key? key,
-    required this.controller,
-  }) : super(key: key);
-
-  final ReviewerSubmitController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.grey[200],
-      child: Card(
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Expanded(
-                    flex: 6,
-                    child: GFTypography(
-                      text: 'Karakter Debitur',
-                      type: GFTypographyType.typo3,
-                      showDivider: false,
-                    ),
-                  ),
-                  Expanded(
-                    child: Obx(
-                      () => Icon(
-                        controller.isKarakterPressed.value == true
-                            ? Icons.check_box
-                            : Icons.close,
-                        color: controller.isKarakterPressed.value == true
-                            ? Colors.green
-                            : Colors.transparent,
-                        size: 30,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 10),
-              Text(
-                'Ini merupakan rekapan karakter debitur, pilih ya atau tidak untuk mengirim response ini ke pemutus',
-                style: Theme.of(context).textTheme.bodySmall?.merge(
-                      const TextStyle(
-                        fontSize: 14,
-                      ),
-                    ),
-              ),
-              const SizedBox(height: 10),
-              GFButton(
-                onPressed: () {
-                  Get.toNamed(Routes.KARAKTER_PRINT,
-                      arguments: controller.insightDebitur.value);
-                },
-                text: 'Lihat Summary Karakter',
-                color: primaryColor,
-                shape: GFButtonShape.pills,
-                icon: const Icon(
-                  Icons.summarize,
-                  size: 18,
-                  color: secondaryColor,
-                ),
-                fullWidthButton: true,
-                size: GFSize.LARGE,
-              ),
-              const SizedBox(height: 10),
-              FormBuilderRadioGroup(
-                name: 'karakter',
-                wrapAlignment: WrapAlignment.center,
-                decoration: InputDecoration(
-                  floatingLabelAlignment: FloatingLabelAlignment.center,
-                  labelStyle: Theme.of(context).textTheme.bodySmall!.merge(
-                        const TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                  border: InputBorder.none,
-                  alignLabelWithHint: true,
-                  label: const Text('Apakah karakter debitur ini layak?'),
-                ),
-                options: const [
-                  FormBuilderFieldOption(
-                    value: true,
-                    child: Text('👍 Ya'),
-                  ),
-                  FormBuilderFieldOption(
-                    value: false,
-                    child: Text('👎 Tidak'),
-                  ),
-                ],
-                onChanged: (value) {
-                  controller.isKarakterPressed.value = true;
-                },
-                validator: FormBuilderValidators.required(),
-              ),
             ],
           ),
         ),
