+++ lib/app/modules/reviewer_submit/widget/reviewer_submit_karakter.dart
@@ -1,5 +1,7 @@
+// 🎯 Dart imports:
+
 // 🐦 Flutter imports:
-import 'package:akm/app/modules/reviewer_submit/widget/components/reviewer_submit_keuangan.dart/reviewer_submit_keuangan_buttons.dart';
+import 'package:akm/app/modules/reviewer_submit/widget/components/reviewer_submit_karakter/reviewer_submit_karakter_button.dart';
 import 'package:flutter/material.dart';
 
 // 📦 Package imports:
@@ -9,20 +11,14 @@ import 'package:getwidget/getwidget.dart';
 // 🌎 Project imports:
 import '../controllers/reviewer_submit_controller.dart';
 
-class ReviewerKeuanganSection extends StatelessWidget {
-  const ReviewerKeuanganSection({
+class ReviewerKarakterSection extends StatelessWidget {
+  const ReviewerKarakterSection({
     Key? key,
     required this.controller,
-    required this.iconNotYet,
-    required this.iconDone,
-    required this.buttonStyle,
     required this.subtitleStyle,
   }) : super(key: key);
 
   final ReviewerSubmitController controller;
-  final Icon iconNotYet;
-  final Icon iconDone;
-  final TextStyle buttonStyle;
   final TextStyle subtitleStyle;
 
   @override
@@ -45,7 +41,7 @@ class ReviewerKeuanganSection extends StatelessWidget {
                   const Expanded(
                     flex: 6,
                     child: GFTypography(
-                      text: 'Keuangan Debitur',
+                      text: 'Karakter Debitur',
                       type: GFTypographyType.typo3,
                       showDivider: false,
                     ),
@@ -53,10 +49,10 @@ class ReviewerKeuanganSection extends StatelessWidget {
                   Expanded(
                     child: Obx(
                       () => Icon(
-                        controller.isKeuanganPressed.value == true
+                        controller.isKarakterPressed.value == true
                             ? Icons.check_box
                             : Icons.close,
-                        color: controller.isKeuanganPressed.value == true
+                        color: controller.isKarakterPressed.value == true
                             ? Colors.green
                             : Colors.transparent,
                         size: 30,
@@ -67,15 +63,12 @@ class ReviewerKeuanganSection extends StatelessWidget {
               ),
               const SizedBox(height: 10),
               Text(
-                'Ini merupakan rekapan keuangan debitur, pilih ya atau tidak untuk mengirim response ini ke pemutus',
+                'Ini merupakan rekapan karakter debitur, pilih ya atau tidak untuk mengirim response ini ke pemutus',
                 style: subtitleStyle,
               ),
               const SizedBox(height: 20),
-              ReviewerSubmitKeuanganButton(
+              ReviewerSubmitKarakterButton(
                 controller: controller,
-                iconDone: iconDone,
-                iconNotYet: iconNotYet,
-                buttonStyle: buttonStyle,
               ),
             ],
           ),
